cleanup testing harnesses
